+++ src/components/table.js
@@ -7,16 +7,15 @@ import storeContext from "../modules/store";
 const Table = () => {
 	const state = useContext(storeContext)[0];
 
-	const { riskLevel, initialIvestement, cones } = state;
+	const { riskLevel, initialIvestement, fee, years, monthlySum, cones } = state;
 	const cone = cones.filter(cone => cone.riskLevel == riskLevel)[0];
-	const fee = 0.01;
 
 	const timeSeries = calculateTimeSeries({
 		mu: cone.mu,
 		sigma: cone.sigma,
-		years: 10,
+		years,
+		monthlySum,
 		initialSum: initialIvestement,
-		monthlySum: 200,
 		fee
 	});